scripts/avail-ext-doc/requirements.txt
Jinja2==2.10
GitPython==2.1.8
PyGithub>=1.36